move code to ui
