+++ src/test/java/tests/ui/BaseUITest.java
@@ -1,4 +1,4 @@
-package tests;
+package tests.ui;
 
 import com.saucelabs.framework.Browser;
 import com.saucelabs.framework.pages.PageObject;
@@ -8,8 +8,10 @@ import org.junit.Before;
 import org.junit.Rule;
 import org.junit.rules.TestName;
 import org.openqa.selenium.remote.RemoteWebDriver;
+import tests.SauceTestWatcher;
+import tests.SessionManager;
 
-public class BaseTest implements SessionManager {
+public class BaseUITest implements SessionManager {
     public SauceSession session;
 
     @Rule
